app/requirements.txt
alabaster==0.7.12
amqp==2.5.2
anyjson==0.3.3
appnope==0.1.0
asgiref==3.2.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.1
billiard==3.6.2.0
bleach==3.1.0
boto3==1.10.34
botocore==1.13.34
cached-property==1.5.1
celery==4.4.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coverage==4.5.4
cryptography==2.8
datedelta==1.3
decorator==4.4.1
defusedxml==0.6.0
Django==3.0
django-admin-interface==0.12.0
django-admin-thumbnails==0.2.4
django-bootstrap-datepicker-plus==3.0.5
django-celery==3.3.1
django-celery-results==1.2.0
django-colorfield==0.1.15
django-countries==5.5
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-extensions==2.2.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-livereload-server==0.3.2
django-money==1.0
django-nose==1.4.6
django-redis==4.11.0
django-storages==1.8
django-timezone-field==3.1
djangorestframework==3.11.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
Faker==3.0.0
freezegun==0.3.14
graphviz==0.13.2
gunicorn==20.0.4
idna==2.8
imagesize==1.2.0
importlib-metadata==1.3.0
ipykernel==5.1.4
ipython==7.11.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.2
Jinja2==2.10.3
jmespath==0.9.4
json2html==1.3.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
kombu==4.6.7
MarkupSafe==1.1.1
mistune==0.8.4
mock==3.0.5
more-itertools==8.0.2
nbconvert==5.6.1
nbformat==5.0.4
nose==1.3.7
notebook==6.0.3
oauthlib==3.1.0
packaging==19.2
pandocfilters==1.4.2
paramiko==2.7.1
parso==0.5.2
pexpect==4.7.0
pg8000==1.13.2
pickleshare==0.7.5
pika==0.13.1
Pillow==6.2.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.2
psutil==5.6.7
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
py-moneyed==0.8.0
pycparser==2.19
pydot==1.4.1
Pygments==2.5.2
PyJWT==1.7.1
PyNaCl==1.3.0
pyparsing==2.4.6
pyrsistent==0.15.6
pytest==5.3.2
pytest-cov==2.8.1
pytest-django==3.8.0
pytest-docker-compose==3.1.2
pytest-env==0.6.2
pytest-pgsql==1.1.1
python-crontab==2.4.0
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.3
PyYAML==3.13
pyzmq==18.1.1
qtconsole==4.6.0
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
roles==0.10
s3transfer==0.2.1
scramp==1.1.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.13.0
snowballstemmer==2.0.0
social-auth-core==3.2.0
soupsieve==1.9.5
Sphinx==2.3.1
sphinx-bootstrap-theme==0.6.5
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
SQLAlchemy==1.3.13
sqlparse==0.3.0
terminado==0.8.3
testing.common.database==2.0.3
testing.postgresql==1.3.0
testpath==0.4.4
text-unidecode==1.3
texttable==1.6.2
tornado==6.0.3
traitlets==4.3.3
urllib3==1.25.7
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
zipp==0.6.0
django-json-widget
django-admin-list-filter-dropdown
django-guardian
colorama
https://github.com/celery/django-celery-beat/zipball/master#egg=django-celery-beat
